backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI API
openai==1.12.0

# Data validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# CORS and security
python-jose[cryptography]==3.3.0

# HTTP client
httpx==0.26.0

# Audio processing
pydub==0.25.1

# Google Calendar Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Yahoo Calendar (CalDAV)
caldav==1.3.9

# Date parsing
python-dateutil==2.8.2
dateparser==1.2.0
